fetch parsing complete.
