+++ xivanalysis/fetch-parse.py
@@ -1,33 +1,74 @@
 import http.client
 import json
 import os
+from enum import Enum
+from selenium import webdriver
+from selenium.webdriver.chrome.options import Options
+from selenium.webdriver.chrome.service import Service
+from webdriver_manager.chrome import ChromeDriverManager
+from selenium.webdriver.common.by import By
+from seleniumwire import webdriver 
+from time import sleep
+import os
 
-class Flag():
-    SOURCEID = 0
-    ENCOUNTER = 1
-    URLREQUEST = 2
-    ROLE = 3
-    JOB = 4
+class Pandaemonium(Enum):
+    P1S =   78
+    P2S =   79
+    P3S =   80
+    P4SP1 = 81
+    P4SP2 = 82
 
-#  ONLY VARIABLES THAT NEED CHANGING in Config.txt  #
-#####################################################
-XIV_PATH = f"{os.path.abspath(os.path.join(os.path.dirname(os.path.realpath(__file__)), os.pardir))}/xivanalysis"
-IGNORE_ACTION = ["attack", "Iron Will", "Provoke", "Hallowed Ground", "Intervention", "Interject", "Shield Wall", "Stronghold", "Last Bastion", "Land Waker", "Dark Force", "Gunmetal Soul"]
-#####################################################
-
-actions = []
-config = []
+class Tank(Enum):
+    PLD = "Paladin"
+    GNB = "Gunbreaker"
+    DRK = "DarkKnight"
+    WAR = "Warrior"
 
 identifiers = {
-  "pld": "Intervene"
+  f"{Tank.PLD.name}": "Intervene"
 }
 
-ROLE = "tank"
-JOB  = "pld"
+JOB     = Tank.PLD
+RAID_TIER = Pandaemonium
+XIV_PATH = f"{os.path.abspath(os.path.join(os.path.dirname(os.path.realpath(__file__)), os.pardir))}/xivanalysis"
+IGNORE_ACTION = ["attack", "Iron Will", "Provoke", "Hallowed Ground", "Intervention", "Interject", "Shield Wall", "Stronghold", "Last Bastion", "Land Waker", "Dark Force", "Gunmetal Soul"]
 
-def single_file(fight_name, fight_url, override = False):
+members = [attr for attr in dir(RAID_TIER) if not callable(getattr(RAID_TIER, attr)) and not attr.startswith("__")]
+actions = []
 
-    if os.path.exists(f"{XIV_PATH}/jobs/{ROLE}/{JOB}/{fight_name}.txt") and not override:
+def load_url(fight_id):
+    options = Options()
+    options.add_argument("--window-size=1920,1080")
+    driver = webdriver.Chrome(service=Service(ChromeDriverManager().install()), options=options)
+
+    driver.get(f"https://www.fflogs.com/zone/rankings/44#boss={fight_id.value}&metric=dps&class=Global&spec={JOB.value}")
+    cookies = driver.find_element(by=By.XPATH ,value="/html/body/div[2]/div/div/div/div[2]/div/button[2]/span")
+    sleep(1)
+    cookies.click()
+    top_player = driver.find_element(by=By.XPATH ,value="/html/body/div[3]/div[3]/div[2]/div[2]/div/div[4]/div[2]/table/tbody/tr[1]/td[2]/div/div[1]/a[1]")
+    sleep(1)
+    top_player.click()
+
+    fflog = driver.current_url[31:].partition("#")
+    fight_url= fflog[0]
+    fight_id = fflog[2][6:].partition("&")[0]
+
+    driver.get(f"https://xivanalysis.com/fflogs/{fight_url}/{fight_id}")
+    sleep(1)
+    top_player = driver.find_element(by=By.XPATH ,value="/html/body/div/div[2]/div[2]/div/div[1]/div/a[1]/span[1]")
+    top_player.click()
+    sleep(2)
+
+    api_url = "Not Found"
+    for request in driver.requests:
+        if request.response:
+            if "https://xivanalysis.com/proxy/fflogs/report/events/" in request.url:
+                api_url = request.url.partition("events/")[2]
+    driver.quit()
+    return api_url
+
+def create_single_file(fight_name, override = False):
+    if os.path.exists(f"{XIV_PATH}/jobs/{Tank.__qualname__}/{JOB.name}/{fight_name}.txt") and not override:
         print(f"{fight_name} file already exists!")
         return
 
@@ -46,16 +87,21 @@ def single_file(fight_name, fight_url, override = False):
         'sec-fetch-site': "same-origin",
         'user-agent': "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/103.0.0.0 Safari/537.36"
         }
-    conn.request("GET", f"/proxy/fflogs/report/events/{fight_url}", payload, headers)
+    url = load_url(RAID_TIER[fight_name])
+    conn.request("GET", f"/proxy/fflogs/report/events/{url}", payload, headers)
 
     result = conn.getresponse()
     data = result.read()
     string = data.decode("utf-8")
 
+    if "429 Too many requests" in string:
+        sleep(5)
+        create_single_file(fight_name, override)
+
     encounter = json.loads(string)
     for entry in encounter['events']:
         try:
-            if entry['ability']['name'] == identifiers[JOB]:
+            if entry['ability']['name'] == identifiers[JOB.name]:
                 sourceID = entry["sourceID"]
                 break
         except:
@@ -74,21 +120,20 @@ def single_file(fight_name, fight_url, override = False):
             actions.append(entry["ability"]["name"])
             actions.append(entry["timestamp"])
 
-    action_output_path = f"{XIV_PATH}/jobs/{ROLE}/{JOB}/{fight_name}.txt"
+    action_output_path = f"{XIV_PATH}/jobs/{Tank.__qualname__}/{JOB.name}/{fight_name}.txt"
     with open(action_output_path, 'w') as f:
         for line in actions:
             f.write(f"{line}\n")
     actions.clear()
 
-def raid_tier():
-    single_file("p1s", "N6nvcKYmQC2d3M4W?start=1994517&end=2406988&translate=true")
-    single_file("p2s", "W6Avqfd1T9mx3FGg?start=450362&end=991804&translate=true")
-    single_file("p3s", "wZKRXbpP7knF1TQY?start=971086&end=1516533&translate=true")
-    single_file("p4sp1", "JfG2yaQPLHYgqR8c?start=216064&end=606434&translate=true")
+def create_raid_tier(force_refresh = False):
+    amount_of_fights = (RAID_TIER[members[-1]].value + 1) - RAID_TIER[members[0]].value
+    for x in range(amount_of_fights):
+        create_single_file(members[x], force_refresh)
 
 def main():
-    #single_file("p1s", "N6nvcKYmQC2d3M4W?start=1994517&end=2406988&translate=true")
-    raid_tier()
+    #create_single_file("P2S")
+    create_raid_tier()
 
 if __name__ == '__main__':
     main()
